requirements.txt
fastapi
uvicorn
numpy
pillow
tensorflow
requests
gdown
scikit-learn
keras